Move S3 connection handling to its own class
This means that the main bucket also benefits from the access denied
`anon=True` fall-back previously only used for the upstream bucket.
This helps with the case where AWS credentials are found, but are
for an account unrelated to either bucket.

Since the main bucket can now fall back to anonymous mode, a check
has been added to the deploy command to improve the UX.

Fixes:
https://github.com/kennethreitz/bob-builder/issues/26#issuecomment-301058112