Remove boto 'set_acl' call after upload
Since it causes access denied errors when the IAM user does not have
`PutBucketAcl` permissions for the S3 bucket, preventing the use of a
more locked down IAM user for deployment.

It seems preferable to leave the bucket ACLs to the bucket owner,
and for `bob deploy` to only upload the file.

Closes W-8134681.